Allow refresh tokens re-authentication
A refresh token is used to generate a new access token. Typically, if the access token has an expiration date, once it expires, the user would have to authenticate again to obtain an access token. With refresh token, this step can be skipped and with a request to the API get a new access token that allows the user to continue accessing the application resources.

Refresh tokens are preferred because credentials authentication should be rate limited (to prevent brute force attacks). Refresh tokens are pretty long and have a high entropy, so brute forcing refresh tokens should be harder.

With this feature restful-client can be configured to return a refresh_token on the authentication process (when configuration value for refresh_token_expiration is different than 0), which can be used later to authenticate using the same auth endpoint.