Optional fact request id
Since facts can be sent to restful client directly without an associated request, the foreign key on facts database is causing those facts to not be persisted on the database.

This comit changes the database request id to be an optional foreign key, so we can store facts without an associated request id.